No more warnings.
git-svn-id: https://wiki.csc.calpoly.edu/svn/AsteroidBlaster@502 342171fe-4ba3-422f-b357-09b6f24955fe

+++ Graphics/Sprite.cpp
@@ -51,7 +51,7 @@ void Sprite::draw() {
    //glDepthFunc(GL_ALWAYS);
    glDepthFunc(GL_LESS);
    double curTime = doubleTime();
-   int curFrame = floor((curTime - startTime) * framesPerSecond);
+   int curFrame = (int) floor((curTime - startTime) * framesPerSecond);
    if (curFrame >= totalFrames && oneShot) {
       shouldRemove = true;
    }
